requirements.txt
asgiref==3.10.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
google-auth==2.41.1
idna==3.10
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
gunicorn
